+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.10.36] = 2025-01-30
+
+### Changed
+
+- Row functions in Admins use HXL tags instead of headers
+- Poverty rate columns updated
+
 ## [0.10.35] = 2025-01-27
 
 ### Changed